refactor: vector search modules
