+++ models/vector/schemes.py
@@ -1,15 +1,12 @@
-"""Vector search model implementation."""
+"""Vector search weighting schemes."""
+
 from collections import defaultdict
-from heapq import nlargest
+from typing import Dict, List, Union, Type
 from math import sqrt, log10
-from typing import List, Dict, Type, Union
 
-import click
-
-from cli_utils import CollectionType, WeightingSchemeClassType
-from datatypes import DocID, Term, PostingList
-from indexes import Index, build_index
 from data_collections import Collection
+from datatypes import DocID, PostingList, Term
+from indexes import Index
 
 
 class WeightingScheme:
@@ -124,73 +121,3 @@ SCHEMES: Dict[str, Type[WeightingScheme]] = {}
 
 for _wcs in (TfIdfSimple, TfIdfComplex):
     SCHEMES[_wcs.name] = _wcs
-
-
-def vector_search(
-    request: str, index: Index, k: int = 10, wcs: Type[WeightingScheme] = None
-) -> List[DocID]:
-    """Perform a vector-space search.
-
-    Parameters
-    ----------
-    request : str
-        A request as a string of words.
-    index : Index
-        A search index.
-    k : int, optional
-        Maximum number of documents to return. Defaults to 10.
-    wcs : class, optional
-        A weighting scheme class. Defaults to `TfIdf`.
-    """
-    if wcs is None:
-        wcs = TfIdfSimple
-
-    scores: Dict[DocID, float] = {doc_id: 0 for doc_id in index.doc_ids}
-    # Weights of request terms
-    wq: List[float] = []
-    w = wcs(index=index, query=list(Collection().tokenize(request)))
-
-    for term_id, term in enumerate(w.query):
-        w_i_q = w.weights[term_id][request] = w.tf(term, request) * w.df(term)
-        wq.append(w_i_q)
-
-        for doc_id in index.postings[term]:
-            w_i_dj = w(term, doc_id)
-            w.weights[term_id][doc_id] = w_i_dj
-            scores[doc_id] += w_i_dj * w_i_q
-
-    norm_q = sum(w_i_q ** 2 for w_i_q in wq)
-
-    for doc_id in index.doc_ids:
-        if scores[doc_id]:
-            scores[doc_id] /= sqrt(w.norm(doc_id)) * sqrt(norm_q) or 1
-
-    top_k: list = nlargest(k, scores.items(), key=lambda item: item[1])
-    return [doc_id for doc_id, _ in top_k]
-
-
-@click.command()
-@click.argument("collection", type=CollectionType())
-@click.argument("query")
-@click.option("--topk", "-k", type=int, default=10)
-@click.option(
-    "--weighting-scheme", "-w", "wcs", type=WeightingSchemeClassType(SCHEMES)
-)
-def cli(
-    collection: Collection, query: str, topk: int, wcs: Type[WeightingScheme]
-):
-    """Test the boolean model on a collection."""
-    index = build_index(collection)
-
-    click.echo(f"Weighting scheme: {wcs.name}")
-
-    click.echo("Query: ", nl=False)
-    click.echo(click.style(query, fg="blue"))
-
-    results = vector_search(query, index, k=topk, wcs=wcs)
-
-    click.echo(click.style(f"Results: {results}", fg="green"))
-
-
-if __name__ == "__main__":
-    cli()
